requirements.txt
absl-py==0.9.0
appdirs==1.4.3
appnope==0.1.0
astor==0.8.1
astroid==2.3.3
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.0
bs4==0.0.1
cachetools==4.1.1
camelot-py==0.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1
colorama==0.4.3
configobj==5.0.6
configparser==4.0.2
cryptography==2.8
cycler==0.10.0
Cython==0.29.21
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
et-xmlfile==1.0.1
etelemetry==0.1.2
filelock==3.0.12
fitz==0.0.1.dev2
gast==0.2.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-resnet==0.1.0
keras-retinanet @ file:///C:/Users/ralha/OneDrive/Desktop/Table_parsing/keras-retinanet
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
neurdflib==5.0.1
nibabel==3.0.2
nipype==1.4.2
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.4.0.40
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pbr==5.4.4
pdf2image==1.13.1
pdfminer.six==20200726
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
progressbar2==3.51.4
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.12.4
prov==1.5.3
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools-windows==2.0.0.2
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.5.2
pylint==2.4.4
PyMuPDF==1.16.11
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyPdf==1.13
PyPDF2==1.26.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2019.3
pywin32==228
pywinpty==0.5.7
pyxnat==1.2.1.0.post3
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.1
QtPy==1.9.0
rdflib==4.2.2
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
sortedcontainers==2.2.2
soupsieve==2.0.1
stevedore==1.32.0
tensorboard==2.1.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
traits==6.0.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
virtualenv==20.0.8
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.8
zipp==2.2.0
